Python script to perform basic operations on IBM Bluemix object storage
One of the most common uses of “Clouds”, is shared or backup storage. A
number of services provide free (limited) storage, and several provide
an easy to use, comfortable interface, such as a folder (subdirectory)
on your desktop where you may drop file to be automatically backed up
(to the cloud service and retrieved or even shared between users), or
web based interface.

Created a utility to provide “Storage as a Service” which:
-Stores a copy of a local file, encrypted, on self created (Cloud)
service. (a small app)
-Retrieves (and decrypt) a remote file
-Lists all files on your service
-Deletes a remote file
-Limits the maximum sizes of remote files to 1 MB, or less, each and no
more than 10 MB, in total.
-Provides an easy to use, “local” interface (terminal based code)